yoga_backend/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
idna==3.10
psycopg2==2.9.10
python-decouple==3.8
requests==2.32.3
sqlparse==0.5.3
urllib3==2.4.0
